+++ lib/testcase/sparql/TestCaseQueryEvaluation.ts
@@ -14,7 +14,7 @@ import type { IFetchOptions, IFetchResponse } from '../../Util';
 import { Util } from '../../Util';
 import type { ITestCaseData } from '../ITestCase';
 import type { ITestCaseHandler } from '../ITestCaseHandler';
-import type { IQueryEngine, IQueryResult, IQueryResultBindings } from './IQueryEngine';
+import type { IQueryEngine, IQueryResult, IQueryResultBindings, IQueryResultBoolean } from './IQueryEngine';
 import type { ITestCaseSparql } from './ITestCaseSparql';
 import { QueryResultBindings } from './QueryResultBindings';
 import { QueryResultBoolean } from './QueryResultBoolean';
@@ -131,9 +131,14 @@ export class TestCaseQueryEvaluationHandler implements ITestCaseHandler<TestCase
    * Parses query results in the DAWG vocabulary.
    * https://www.w3.org/2001/sw/DataAccess/tests/test-dawg.n3
    * @param {Quad[]} quads An array of quads.
-   * @return {Promise<IQueryResultBindings>} A promise resolving to a bindings results object.
+   * @return {Promise<IQueryResultBindings | IQueryResultBoolean>} A promise resolving to a bindings or boolean results object.
    */
-  public static async parseDawgResultSet(quads: RDF.Quad[]): Promise<IQueryResultBindings> {
+  public static async parseDawgResultSet(quads: RDF.Quad[]): Promise<IQueryResultBindings | IQueryResultBoolean> {
+    // Check for a boolean result (ASK query)
+    if (quads[1] && quads[1].predicate.value === 'http://www.w3.org/2001/sw/DataAccess/tests/result-set#boolean') {
+      return new QueryResultBoolean(quads[1].object.value === 'true');
+    }
+
     // Construct resources for easier interpretation of the bindings
     const objectLoader = new RdfObjectLoader({
       context: {
@@ -302,12 +307,12 @@ export class TestCaseQueryEvaluationHandler implements ITestCaseHandler<TestCase
       if (graphData.property.graph) {
         queryDataLinks.push({
           dataUri: graphData.property.graph.value,
-          dataGraph: DF.namedNode(Util.normalizeBaseUrl(graphData.property.label.value)),
+          dataGraph: DF.namedNode(graphData.property.label.value),
         });
       } else {
         queryDataLinks.push({
           dataUri: graphData.value,
-          dataGraph: DF.namedNode(Util.normalizeBaseUrl(graphData.value)),
+          dataGraph: DF.namedNode(graphData.value),
         });
       }
     }
@@ -322,7 +327,7 @@ export class TestCaseQueryEvaluationHandler implements ITestCaseHandler<TestCase
   public static async resolveQueryDataLinks(queryDataLinks: IQueryDataLink[], options?: IFetchOptions): Promise<RDF.Quad[]> {
     let queryData: RDF.Quad[] = [];
     for (const queryDataLink of queryDataLinks) {
-      let queryDataThis: RDF.Quad[] = await arrayifyStream((await Util.fetchRdf(queryDataLink.dataUri, { ...options, normalizeUrl: true }))[1]);
+      let queryDataThis: RDF.Quad[] = await arrayifyStream((await Util.fetchRdf(queryDataLink.dataUri, options))[1]);
       if (queryDataLink.dataGraph) {
         queryDataThis = queryDataThis.map(quad => mapTerms(quad, (value: RDF.Term, key: QuadTermName) => key === 'graph' ? queryDataLink.dataGraph : value));
       }
@@ -360,7 +365,7 @@ export class TestCaseQueryEvaluationHandler implements ITestCaseHandler<TestCase
     return new TestCaseQueryEvaluation(
       testCaseData,
       {
-        baseIRI: Util.normalizeBaseUrl(action.property.query.value),
+        baseIRI: action.property.query.value,
         queryDataLinks,
         laxCardinality,
         queryData,
@@ -418,7 +423,17 @@ export class TestCaseQueryEvaluation implements ITestCaseSparql {
   }
 
   public async test(engine: IQueryEngine, injectArguments: any): Promise<void> {
-    const result: IQueryResult = await engine.query(this.queryData, this.queryString, { baseIRI: this.baseIRI, ...injectArguments });
+    const result: IQueryResult = await engine.query(
+      this.queryData,
+      this.queryString,
+      {
+        baseIRI: this.baseIRI,
+        checkOrder: this.queryResult.type === 'bindings' ?
+          this.queryResult.checkOrder :
+          false,
+        ...injectArguments,
+      },
+    );
     if (!this.queryResult.equals(result, this.laxCardinality)) {
       throw new ErrorTest(`Invalid query evaluation